requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0
flask>=3.0.0
flask-cors>=4.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
gunicorn>=21.2.0
